+++ crates/stackable-versioned-macros/src/attrs/common/container.rs
@@ -12,15 +12,19 @@ use k8s_version::Version;
 /// Currently supported attributes are:
 ///
 /// - `version`, which can occur one or more times. See [`VersionAttributes`].
-/// - `options`, which allow further customization of the generated code. See [`ContainerOptions`].
+/// - `options`, which allow further customization of the generated code.
+///    See [`ContainerAttributes`].
 #[derive(Debug, FromMeta)]
 #[darling(and_then = ContainerAttributes::validate)]
 pub(crate) struct ContainerAttributes {
     #[darling(multiple, rename = "version")]
     pub(crate) versions: SpannedValue<Vec<VersionAttributes>>,
 
-    #[darling(default)]
-    pub(crate) options: ContainerOptions,
+    #[darling(rename = "k8s")]
+    pub(crate) kubernetes_attrs: Option<KubernetesAttributes>,
+
+    #[darling(default, rename = "options")]
+    pub(crate) common_option_attrs: OptionAttributes,
 }
 
 impl ContainerAttributes {
@@ -43,7 +47,7 @@ impl ContainerAttributes {
 
         // Ensure that versions are defined in sorted (ascending) order to keep
         // code consistent.
-        if !self.options.allow_unsorted.is_present() {
+        if !self.common_option_attrs.allow_unsorted.is_present() {
             let original = self.versions.deref().clone();
             self.versions
                 .sort_by(|lhs, rhs| lhs.name.partial_cmp(&rhs.name).unwrap_or(Ordering::Equal));
@@ -71,18 +75,26 @@ impl ContainerAttributes {
         // place.
 
         // Ensure every version is unique and isn't declared multiple times.
-        let duplicates = self
+        let duplicate_versions = self
             .versions
             .iter()
             .duplicates_by(|e| e.name)
             .map(|e| e.name)
             .join(", ");
 
-        if !duplicates.is_empty() {
+        if !duplicate_versions.is_empty() {
             return Err(Error::custom(format!(
-                "attribute macro `#[versioned()]` contains duplicate versions: {duplicates}",
+                "attribute macro `#[versioned()]` contains duplicate versions: {duplicate_versions}",
             ))
             .with_span(&self.versions.span()));
+        }
+
+        // Ensure that the 'k8s' feature is enabled when the 'k8s()'
+        // attribute is used.
+        if self.kubernetes_attrs.is_some() && cfg!(not(feature = "k8s")) {
+            return Err(Error::custom(
+                "the `#[versioned(k8s())]` attribute can only be used when the `k8s` feature is enabled",
+            ));
         }
 
         Ok(self)
@@ -101,29 +113,59 @@ impl ContainerAttributes {
 pub(crate) struct VersionAttributes {
     pub(crate) deprecated: Flag,
     pub(crate) name: Version,
-    pub(crate) skip: Option<SkipOptions>,
+    pub(crate) skip: Option<CommonSkipAttributes>,
     pub(crate) doc: Option<String>,
 }
 
-/// This struct contains supported container options.
+/// This struct contains supported option attributes.
 ///
-/// Supported options are:
+/// Supported attributes are:
 ///
 /// - `allow_unsorted`, which allows declaring versions in unsorted order,
 ///   instead of enforcing ascending order.
 /// - `skip` option to skip generating various pieces of code.
 #[derive(Clone, Debug, Default, FromMeta)]
-pub(crate) struct ContainerOptions {
+pub(crate) struct OptionAttributes {
     pub(crate) allow_unsorted: Flag,
-    pub(crate) skip: Option<SkipOptions>,
+    pub(crate) skip: Option<CommonSkipAttributes>,
 }
 
-/// This struct contains supported skip options.
+/// This struct contains supported Kubernetes attributes.
 ///
-/// Supported options are:
+/// Supported attributes are:
+///
+/// - `skip`, which controls skipping parts of the generation.
+/// - `kind`, which allows overwriting the kind field of the CRD. This defaults
+///    to the struct name (without the 'Spec' suffix).
+/// - `group`, which sets the CRD group, usually the domain of the company.
+#[derive(Clone, Debug, FromMeta)]
+pub(crate) struct KubernetesAttributes {
+    pub(crate) skip: Option<KubernetesSkipAttributes>,
+    pub(crate) kind: Option<String>,
+    pub(crate) group: String,
+}
+
+/// This struct contains supported kubernetes skip attributes.
+///
+/// Supported attributes are:
+///
+/// - `merged_crd` flag, which skips generating the `crd()` and `merged_crd()`
+///   functions are generated.
+#[derive(Clone, Debug, FromMeta)]
+pub(crate) struct KubernetesSkipAttributes {
+    /// Whether the `crd()` and `merged_crd()` generation should be skipped for
+    /// this container.
+    pub(crate) merged_crd: Flag,
+}
+
+/// This struct contains supported common skip attributes.
+///
+/// Supported attributes are:
 ///
 /// - `from` flag, which skips generating [`From`] implementations when provided.
 #[derive(Clone, Debug, Default, FromMeta)]
-pub(crate) struct SkipOptions {
+pub(crate) struct CommonSkipAttributes {
+    /// Whether the [`From`] implementation generation should be skipped for all
+    /// versions of this container.
     pub(crate) from: Flag,
 }